Move `OperationTarget` to `OutlineTree`
